pyproject.toml
[project]
name = "thraeka"
version = "0.1.0"
description = ""
authors = [
    {name = "Travis Dang"}
]
requires-python = "3.12.0"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "django-extensions (>=3.2.3,<4.0.0)",
    "pymupdf (>=1.25.3,<2.0.0)",
    "openai (>=1.65.3,<2.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "django-filter (>=25.1,<26.0)"
]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
black = "^25.1.0"
mypy = "^1.15.0"
isort = "^6.0.0"
pytest = "^8.3.4"
pytest-django = "^4.10.0"
pre-commit = "^4.1.0"
pylint = "^3.3.4"
